backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.18
redis==5.2.1
slowapi==0.1.9
httpx==0.28.1
email-validator==2.2.0
google-auth==2.36.0
pypdf==5.1.0
python-docx==1.1.2
scikit-learn==1.5.2
numpy==2.0.2
boto3==1.35.76
pytest==8.3.4
pytest-asyncio==0.24.0
aiosmtplib==3.0.2
jinja2==3.1.4

# Optional NLP: uncomment for spaCy-enhanced extraction (large download)
# spacy==3.7.5
